:recycle: refactor migrate issue report
