+++ api/lib/routes.js
@@ -6,7 +6,6 @@ import * as certificationCenterInvitations from './application/certification-cen
 import * as certificationCenterMemberships from './application/certification-center-memberships/index.js';
 import * as certificationCenters from './application/certification-centers/index.js';
 import * as certificationCourses from './application/certification-courses/index.js';
-import * as certificationIssueReports from './application/certification-issue-reports/index.js';
 import * as certificationPointOfContacts from './application/certification-point-of-contacts/index.js';
 import * as certifications from './application/certifications/index.js';
 import * as frameworks from './application/frameworks/index.js';
@@ -31,7 +30,6 @@ const routes = [
   certificationCenterMemberships,
   certificationCourses,
   certificationPointOfContacts,
-  certificationIssueReports,
   certifications,
   healthcheck,
   memberships,
