+++ api/src/certification/session-management/application/certification-issue-report-route.js
@@ -1,7 +1,7 @@
 import Joi from 'joi';
 
-import { securityPreHandlers } from '../../../src/shared/application/security-pre-handlers.js';
-import { identifiersType } from '../../../src/shared/domain/types/identifiers-type.js';
+import { securityPreHandlers } from '../../../shared/application/security-pre-handlers.js';
+import { identifiersType } from '../../../shared/domain/types/identifiers-type.js';
 import { certificationIssueReportController } from './certification-issue-report-controller.js';
 
 const register = async function (server) {
@@ -21,7 +21,7 @@ const register = async function (server) {
             id: identifiersType.certificationIssueReportId,
           }),
         },
-        handler: certificationIssueReportController.deleteCertificationIssueReport,
+        handler: certificationIssueReportController.deleteCertification,
         tags: ['api', 'certification-issue-reports'],
         notes: [
           '- **Cette route est restreinte aux utilisateurs qui sont membres du centre de certification**\n',
