+++ api/src/certification/session-management/domain/usecases/delete-certification-issue-report.js
@@ -1,4 +1,4 @@
-import { ForbiddenAccess } from '../../../src/shared/domain/errors.js';
+import { ForbiddenAccess } from '../../../../shared/domain/errors.js';
 
 const deleteCertificationIssueReport = async function ({
   certificationIssueReportId,
