+++ api/src/certification/session-management/infrastructure/repositories/index.js
@@ -83,10 +83,10 @@ const repositoriesWithoutInjectedDependencies = {
   jurySessionRepository,
   sessionForInvigilatorKitRepository,
   issueReportCategoryRepository,
-  certificationIssueReportRepository,
   sessionRepository,
   sessionForSupervisingRepository,
   sessionJuryCommentRepository,
+  certificationIssueReportRepository,
   supervisorAccessRepository,
   certificationReportRepository,
   v3CertificationCourseDetailsForAdministrationRepository,
@@ -113,6 +113,7 @@ export {
   assessmentRepository,
   assessmentResultRepository,
   certificationChallengeRepository,
+  certificationIssueReportRepository,
   challengeRepository,
   competenceMarkRepository,
   cpfExportRepository,
