+++ api/tests/certification/session-management/unit/application/certification-issue-reports-route_test.js
@@ -1,7 +1,7 @@
-import { certificationIssueReportController } from '../../../../lib/application/certification-issue-reports/certification-issue-report-controller.js';
-import * as moduleUnderTest from '../../../../lib/application/certification-issue-reports/index.js';
-import { securityPreHandlers } from '../../../../src/shared/application/security-pre-handlers.js';
-import { expect, HttpTestServer, sinon } from '../../../test-helper.js';
+import { certificationIssueReportController } from '../../../../../src/certification/session-management/application/certification-issue-report-controller.js';
+import * as moduleUnderTest from '../../../../../src/certification/session-management/application/certification-issue-report-route.js';
+import { securityPreHandlers } from '../../../../../src/shared/application/security-pre-handlers.js';
+import { expect, HttpTestServer, sinon } from '../../../../test-helper.js';
 
 describe('Unit | Application | Certifications Issue Report | Route', function () {
   describe('DELETE /api/certification-issue-reports/{id}', function () {
@@ -10,7 +10,7 @@ describe('Unit | Application | Certifications Issue Report | Route', function ()
       sinon
         .stub(securityPreHandlers, 'checkUserIsMemberOfCertificationCenterSessionFromCertificationIssueReportId')
         .callsFake((request, h) => h.response(true));
-      sinon.stub(certificationIssueReportController, 'deleteCertificationIssueReport').returns('ok');
+      sinon.stub(certificationIssueReportController, 'deleteCertification').returns('ok');
 
       const httpTestServer = new HttpTestServer();
       await httpTestServer.register(moduleUnderTest);
